requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv export --frozen --no-hashes --no-emit-project --all-groups --all-extras --output-file=requirements-dev.txt
adbc-driver-manager==1.4.0 ; python_full_version < '4.0'
adbc-driver-postgresql==1.4.0 ; python_full_version < '4.0'
aiohappyeyeballs==2.4.4 ; python_full_version < '4.0'
aiohttp==3.11.11 ; python_full_version < '4.0'
aiosignal==1.3.2 ; python_full_version < '4.0'
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4 ; python_full_version < '4.0'
appnope==0.1.4 ; sys_platform == 'darwin'
asn1crypto==1.5.1 ; python_full_version < '4.0'
asttokens==3.0.0
async-timeout==5.0.1 ; python_full_version < '3.11'
atpublic==5.1
attrs==25.1.0
backports-tarfile==1.2.0 ; python_full_version < '3.12'
beartype==0.19.0
beniget==0.4.2.post1 ; sys_platform == 'darwin'
black==25.1.0
blackdoc==0.3.9
cachetools==5.5.1 ; python_full_version < '4.0'
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
cityhash==0.4.7 ; python_full_version < '4.0'
click==8.1.8
cloudpickle==3.1.1 ; python_full_version < '4.0'
codespell==2.4.1
colorama==0.4.6
comm==0.2.2
connectorx==0.4.1 ; python_full_version < '4.0'
coverage==7.6.10
cryptography==44.0.0 ; python_full_version < '4.0' or sys_platform == 'linux'
dask==2025.1.0 ; python_full_version < '4.0'
datafusion==43.1.0
debugpy==1.8.12
decorator==5.1.1
distlib==0.3.9
duckdb==1.1.3
exceptiongroup==1.2.2 ; python_full_version < '3.11'
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
frozenlist==1.5.0 ; python_full_version < '4.0'
fsspec==2024.12.0 ; python_full_version < '4.0'
gast==0.6.0 ; sys_platform == 'darwin'
gcsfs==2024.12.0 ; python_full_version < '4.0'
geoarrow-types==0.2.0 ; python_full_version < '4.0'
git-cliff==2.8.0
google-api-core==2.24.1 ; python_full_version < '4.0'
google-auth==2.38.0 ; python_full_version < '4.0'
google-auth-oauthlib==1.2.1 ; python_full_version < '4.0'
google-cloud-core==2.4.1 ; python_full_version < '4.0'
google-cloud-storage==3.0.0 ; python_full_version < '4.0'
google-crc32c==1.6.0 ; python_full_version < '4.0'
google-resumable-media==2.7.2 ; python_full_version < '4.0'
googleapis-common-protos==1.66.0 ; python_full_version < '4.0'
greenlet==3.1.1 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
griffe==1.5.6
h11==0.14.0
hatch==1.14.0
hatchling==1.27.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0 ; python_full_version < '4.0'
hyperlink==21.0.0
hypothesis==6.124.9
identify==2.6.6
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.1.0
jedi==0.19.2
jeepney==0.8.0 ; sys_platform == 'linux'
jinja2==3.1.5 ; python_full_version < '4.0'
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
keyring==25.6.0
locket==1.0.0 ; python_full_version < '4.0'
lz4==4.4.3
markdown-it-py==3.0.0
markupsafe==3.0.2 ; python_full_version < '4.0'
matplotlib-inline==0.1.7
maturin==1.8.1
mdurl==0.1.2
more-itertools==10.6.0
multidict==6.1.0 ; python_full_version < '4.0'
mypy-extensions==1.0.0
nbclient==0.10.2
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.2.2
nvidia-nccl-cu12==2.25.1 ; platform_machine != 'aarch64' and sys_platform == 'linux'
oauthlib==3.2.2 ; python_full_version < '4.0'
packaging==24.2
pandas==2.2.3 ; python_full_version < '4.0'
parso==0.8.4
parsy==2.1
partd==1.4.2 ; python_full_version < '4.0'
pathspec==0.12.1
pexpect==4.9.0
pins==0.8.7 ; python_full_version < '4.0'
pip==25.0
platformdirs==4.3.6
pluggy==1.5.0
plum-dispatch==2.5.7
ply==3.11 ; sys_platform == 'darwin'
pre-commit==4.1.0
prompt-toolkit==3.0.50
propcache==0.2.1 ; python_full_version < '4.0'
proto-plus==1.26.0 ; python_full_version < '4.0'
protobuf==5.29.3 ; python_full_version < '4.0'
psutil==6.1.1
psycopg2-binary==2.9.10 ; python_full_version < '4.0'
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.0
pyarrow-hotfix==0.6 ; python_full_version < '4.0'
pyasn1==0.6.1 ; python_full_version < '4.0'
pyasn1-modules==0.4.1 ; python_full_version < '4.0'
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyjwt==2.10.1 ; python_full_version < '4.0'
pyopenssl==24.3.0 ; python_full_version < '4.0'
pytest==8.3.4
pytest-codspeed==3.2.0
pytest-cov==6.0.0
pytest-mock==3.14.0 ; python_full_version < '4.0'
pytest-snapshot==0.9.0
python-dateutil==2.9.0.post0
pythran==0.17.0 ; sys_platform == 'darwin'
pytz==2025.1
pywin32==308 ; platform_python_implementation != 'PyPy' and sys_platform == 'win32'
pywin32-ctypes==0.2.3 ; sys_platform == 'win32'
pyyaml==6.0.2
pyzmq==26.2.1
quartodoc==0.9.1
quickgrove==0.1.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0 ; python_full_version < '4.0'
rich==13.9.4
rpds-py==0.22.3
rsa==4.9 ; python_full_version < '4.0'
ruff==0.9.4
scikit-learn==1.6.1
scipy==1.15.1
secretstorage==3.3.3 ; sys_platform == 'linux'
setuptools==75.8.0 ; sys_platform == 'darwin'
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.13.2 ; python_full_version < '4.0'
sortedcontainers==2.4.0
sphobjinv==2.3.1.2
sqlalchemy==2.0.37 ; python_full_version < '4.0'
sqlglot==25.20.2
stack-data==0.6.3
structlog==25.1.0 ; python_full_version < '4.0'
tabulate==0.9.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.2.1
tomli-w==1.2.0
tomlkit==0.13.2
toolz==0.12.1
tornado==6.4.2
traitlets==5.14.3
trino==0.332.0
trove-classifiers==2025.1.15.22
typing-extensions==4.12.2
tzdata==2025.1 ; python_full_version < '4.0' or sys_platform == 'win32'
tzlocal==5.2
urllib3==2.3.0
userpath==1.9.2
uv==0.5.26
vendoring==1.2.0
virtualenv==20.29.1
watchdog==6.0.0
wcwidth==0.2.13
xgboost==2.1.3
xxhash==3.5.0 ; python_full_version < '4.0'
yarl==1.18.3 ; python_full_version < '4.0'
zipp==3.21.0
zstandard==0.23.0
